APP_AUTH for player logins, fixes game status bug
- APP_ADMIN_AUTH changed to APP_AUTH, targets both the player and admin authentication method. Still defaults to adldap

- Fixes Status bug
Players that submitted an indicment with setting an Evidence room choice resulted in a NULL evidence_selected_at value. This caused the status view to break. evidence_selected_at is set to NOW() when the indictment is made, ONLY if the value is currently NULL.

Additionally, check for a null evidence_selected_at before rendering to prevent failure.